Implemented UUID in stats to distinguish between servers if they rename or change IP address. Added lightning strike effect when a player is killed by a player. (optional) When a player respawns, they can't be killed for x seconds. (optional) Reputation anti-farm (player's reputation can't change too often. (optional) Changed '/pk' command to respect whether or not 'PK Only'-mode is enabled. Miscellaneous fixes here and there. Slight code tidying. Slightly altered the buff values. Not overpowered, but slightly more. Changed the default configuration to have everything disabled and show no messages. Fixed combat to not fire (enter/exit) when duration is set to 0. Added an auto-updater. (optional) Updates require a manual reload or restart. Fixed problem with p.showPlayer(pp); and p.hidePlayer(pp); for VanishNoPacket. Added '/pks debug' command and 'pks.debug' permission node to see debug info. Added 'pks.override.combat' permission node to skip enter-combat checks. Added 'pks.reload.playerdata' permission node.. Forgot that one. Changed playerOpt to playerData to leave room for future additions.

+++ PKSystem/src/com/scizzr/bukkit/plugins/pksystem/listeners/Entities.java
@@ -19,6 +19,7 @@ import com.scizzr.bukkit.plugins.pksystem.managers.Manager;
 import com.scizzr.bukkit.plugins.pksystem.util.MoreMath;
 import com.scizzr.bukkit.plugins.pksystem.util.MoreString;
 import com.scizzr.bukkit.plugins.pksystem.util.TombStone;
+import com.scizzr.bukkit.plugins.pksystem.util.Vault;
 
 public class Entities implements Listener {
     Main plugin;
@@ -52,20 +53,22 @@ public class Entities implements Listener {
             
             if (pDef == pAtt) { pAtt.sendMessage(Main.prefix + "You can't harm yourself"); e.setCancelled(true); return; }
             
-            if (Config.combSpawnEnabled == true && Manager.isRespawn(pDef) == true) {
-                pAtt.sendMessage(Main.prefix + "You can't attack " + pDef.getName() + " yet. They just spawned.");
-                e.setCancelled(true); return;
-            }
-            
-            if (Manager.getIndex(Manager.getPoints(pDef)) >= 0) {
-                if (Manager.isPK(pAtt) == false && !Manager.isCombat(pDef) && Config.combPkOnly == true) {
-                    pAtt.sendMessage(Main.prefix + "You must enter PK mode to kill players");
+            if (!Vault.hasPermission(pAtt, "override.combat")) {
+                if (Config.combSpawnEnabled == true && Manager.isRespawn(pDef) == true) {
+                    pAtt.sendMessage(Main.prefix + "You can't attack " + pDef.getName() + " yet. They just spawned.");
                     e.setCancelled(true); return;
                 }
                 
-                if (Config.combNoobEnabled && (pDef.getLevel() < Config.combNoobLevel && Manager.isCombat(pDef) == false)) {
-                    pAtt.sendMessage(Main.prefix + "You can't kill good players under level " + String.valueOf(Config.combNoobLevel));
-                    e.setCancelled(true); return;
+                if (Manager.getIndex(Manager.getPoints(pDef)) >= 0) {
+                    if (Manager.isPK(pAtt) == false && !Manager.isCombat(pDef) && Config.combPkOnly == true) {
+                        pAtt.sendMessage(Main.prefix + "You must enter PK mode to kill players");
+                        e.setCancelled(true); return;
+                    }
+                    
+                    if (Config.combNoobEnabled && (pDef.getLevel() < Config.combNoobLevel && Manager.isCombat(pDef) == false)) {
+                        pAtt.sendMessage(Main.prefix + "You can't kill good players under level " + String.valueOf(Config.combNoobLevel));
+                        e.setCancelled(true); return;
+                    }
                 }
             }
